requirements.txt
fastapi==0.115.11
pydantic==2.10.6
sqlalchemy==2.0.39
uvicorn==0.34.0
PyJWT==2.10.1
passlib[bcrypt]==1.7.4